tests/requirements_universal.txt
telethon>=1.34.0
clickhouse-driver>=0.2.6
python-dotenv>=1.0.0
